+++ src/components/ProviderInitializer/ProviderInitializer.tsx
@@ -32,10 +32,10 @@ import {
   setAccountLoadingError,
   setLedgerAccount,
   setWalletLogin,
-  setChainID,
   setTokenLogin,
   setIsWalletConnectV2Initialized,
-  setAddress
+  setAddress,
+  updateNetworkConfig
 } from 'reduxStore/slices';
 import { decodeNativeAuthToken } from 'services/nativeAuth/helpers';
 import { LoginMethodsEnum } from 'types/enums.types';
@@ -92,7 +92,7 @@ export function ProviderInitializer() {
   });
 
   useEffect(() => {
-    refreshChainID();
+    refreshNetworkConfig();
   }, [network]);
 
   useEffect(() => {
@@ -112,11 +112,17 @@ export function ProviderInitializer() {
     setLedgerAccountInfo();
   }, [ledgerAccount, isLoggedIn, ledgerData]);
 
-  async function refreshChainID() {
+  async function refreshNetworkConfig() {
     try {
       const networkConfig = await getNetworkConfigFromApi();
+
       if (networkConfig) {
-        dispatch(setChainID(networkConfig.erd_chain_id));
+        dispatch(
+          updateNetworkConfig({
+            chainId: networkConfig.erd_chain_id,
+            roundDuration: networkConfig.erd_round_duration
+          })
+        );
       }
     } catch (err) {
       console.error('failed refreshing chainId ', err);