+++ src/hooks/transactions/batch/tracker/useCheckHangingBatchesFallback.ts
@@ -1,13 +1,11 @@
 import { useCallback, useEffect } from 'react';
-import {
-  AVERAGE_TX_DURATION_MS,
-  TRANSACTIONS_STATUS_DROP_INTERVAL_MS
-} from 'constants/transactionStatus';
-import { extractSessionId } from 'hooks/transactions/helpers/extractSessionId';
-import { timestampIsOlderThan } from 'hooks/transactions/helpers/timestampIsOlderThan';
+import { TRANSACTIONS_STATUS_DROP_INTERVAL_MS } from 'constants/transactionStatus';
 import { removeBatchTransactions } from 'services/transactions';
 import { getTransactionsStatus } from 'utils/transactions/batch/getTransactionsStatus';
 import { sequentialToFlatArray } from 'utils/transactions/batch/sequentialToFlatArray';
+import { extractSessionId } from '../../helpers/extractSessionId';
+import { timestampIsOlderThan } from '../../helpers/timestampIsOlderThan';
+import { useGetPollingInterval } from '../../useGetPollingInterval';
 import { useGetBatches } from '../useGetBatches';
 import { useUpdateBatch } from './useUpdateBatch';
 
@@ -20,6 +18,7 @@ export const useCheckHangingBatchesFallback = (props?: {
   onFail?: (sessionId: string | null, errorMessage?: string) => void;
 }) => {
   const { batchTransactionsArray } = useGetBatches();
+  const pollingInterval = useGetPollingInterval();
   const updateBatch = useUpdateBatch();
   const onSuccess = props?.onSuccess;
   const onFail = props?.onFail;
@@ -69,7 +68,7 @@ export const useCheckHangingBatchesFallback = (props?: {
   useEffect(() => {
     const interval = setInterval(async () => {
       checkHangingBatches();
-    }, AVERAGE_TX_DURATION_MS);
+    }, pollingInterval);
 
     return () => clearInterval(interval);
   }, [checkHangingBatches]);