+++ src/hooks/transactions/useGetPollingInterval.ts
@@ -1,0 +1,16 @@
+import {
+  CROSS_SHARD_ROUNDS,
+  TRANSACTIONS_STATUS_POLLING_INTERVAL_MS
+} from 'constants/transactionStatus';
+import { useSelector } from 'reduxStore/DappProviderContext';
+import { roundDurationSelectorSelector } from 'reduxStore/selectors';
+
+export const useGetPollingInterval = () => {
+  const roundDuration = useSelector(roundDurationSelectorSelector);
+
+  if (!roundDuration) {
+    return TRANSACTIONS_STATUS_POLLING_INTERVAL_MS;
+  }
+
+  return (roundDuration / 2) * CROSS_SHARD_ROUNDS;
+};