+++ src/hooks/transactions/useTransactionsTracker.ts
@@ -1,11 +1,13 @@
 import { useEffect } from 'react';
 import { getTransactionsByHashes as defaultGetTxByHash } from 'apiCalls/transactions';
-import { useCheckTransactionStatus, useRegisterWebsocketListener } from 'hooks';
 import { TransactionsTrackerType } from 'types/transactionsTracker.types';
+import { useRegisterWebsocketListener } from '../websocketListener';
+import { useCheckTransactionStatus } from './useCheckTransactionStatus';
+import { useGetPollingInterval } from './useGetPollingInterval';
 
 export function useTransactionsTracker(props?: TransactionsTrackerType) {
   const checkTransactionStatus = useCheckTransactionStatus();
-
+  const pollingInterval = useGetPollingInterval();
   const getTransactionsByHash =
     props?.getTransactionsByHash ?? defaultGetTxByHash;
 
@@ -20,7 +22,8 @@ export function useTransactionsTracker(props?: TransactionsTrackerType) {
   useRegisterWebsocketListener(onMessage);
 
   useEffect(() => {
-    const interval = setInterval(onMessage, 30000);
+    const interval = setInterval(onMessage, pollingInterval);
+
     return () => {
       clearInterval(interval);
     };