+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added dynamic transaction polling based on network round duration](https://github.com/multiversx/mx-sdk-dapp/pull/1201)
+
 ## [[v2.33.6]](https://github.com/multiversx/mx-sdk-dapp/pull/1200)] - 2024-07-03
 
 - [Update WalletConnect, disconnect default session (QR-code) on existing pairing](https://github.com/multiversx/mx-sdk-dapp/pull/1199)